--- src/Engine/PostProcessing/GlowFilter.cpp
@@ -1,30 +1,0 @@
-#include "GlowFilter.hpp"
-
-#include <Video/Shader/Shader.hpp>
-#include <Video/Shader/ShaderProgram.hpp>
-#include "../Manager/Managers.hpp"
-#include "../Manager/ResourceManager.hpp"
-#include "Post.vert.hpp"
-#include "PostGlow.frag.hpp"
-
-using namespace Video;
-
-GlowFilter::GlowFilter() {
-    vertexShader = Managers().resourceManager->CreateShader(POST_VERT, POST_VERT_LENGTH, GL_VERTEX_SHADER);
-    fragmentShader = Managers().resourceManager->CreateShader(POSTGLOW_FRAG, POSTGLOW_FRAG_LENGTH, GL_FRAGMENT_SHADER);
-    shaderProgram = Managers().resourceManager->CreateShaderProgram({ vertexShader, fragmentShader });
-}
-
-GlowFilter::~GlowFilter() {
-    Managers().resourceManager->FreeShaderProgram(shaderProgram);
-    Managers().resourceManager->FreeShader(vertexShader);
-    Managers().resourceManager->FreeShader(fragmentShader);
-}
-
-ShaderProgram* GlowFilter::GetShaderProgram() const {
-    return shaderProgram;
-}
-
-void GlowFilter::SetUniforms() {
-    
-}
