+++ src/Video/PostProcessing/FXAAFilter.cpp
@@ -1,27 +1,25 @@
 #include "FXAAFilter.hpp"
 
-#include <Video/Shader/Shader.hpp>
-#include <Video/Shader/ShaderProgram.hpp>
-#include "../Manager/Managers.hpp"
-#include "../Manager/ResourceManager.hpp"
+#include "../Shader/Shader.hpp"
+#include "../Shader/ShaderProgram.hpp"
 #include "Post.vert.hpp"
 #include "PostFXAA.frag.hpp"
 
 using namespace Video;
 
 FXAAFilter::FXAAFilter() {
-    vertexShader = Managers().resourceManager->CreateShader(POST_VERT, POST_VERT_LENGTH, GL_VERTEX_SHADER);
-    fragmentShader = Managers().resourceManager->CreateShader(POSTFXAA_FRAG, POSTFXAA_FRAG_LENGTH, GL_FRAGMENT_SHADER);
-    shaderProgram = Managers().resourceManager->CreateShaderProgram({ vertexShader, fragmentShader });
+    Shader* vertexShader = new Shader(POST_VERT, POST_VERT_LENGTH, GL_VERTEX_SHADER);
+    Shader* fragmentShader = new Shader(POSTFXAA_FRAG, POSTFXAA_FRAG_LENGTH, GL_FRAGMENT_SHADER);
+    shaderProgram = new ShaderProgram({ vertexShader, fragmentShader });
+    delete vertexShader;
+    delete fragmentShader;
     
     screenSizeLocation = shaderProgram->GetUniformLocation("screenSize");
     brightnessLocation = shaderProgram->GetUniformLocation("brightness");
 }
 
 FXAAFilter::~FXAAFilter() {
-    Managers().resourceManager->FreeShaderProgram(shaderProgram);
-    Managers().resourceManager->FreeShader(vertexShader);
-    Managers().resourceManager->FreeShader(fragmentShader);
+    delete shaderProgram;
 }
 
 ShaderProgram* FXAAFilter::GetShaderProgram() const {
